+++ packages/umi-preset-react-native/src/plugins/generateFiles/react-native/routes/index.ts
@@ -1,5 +1,6 @@
 import { IApi } from 'umi';
-import { createFormatter, routesToJSON } from '../../../utils';
+import { createFormatter } from '../../../utils';
+import routesToJSON from './routesToJSON';
 
 const ROUTES_TPL = `import {ApplyPluginsType, dynamic} from 'umi';
 import Multibundle from 'umi-react-native-multibundle';
