+++ packages/umi-preset-react-native/src/plugins/generators/rn/getBundleNameFrom.ts
@@ -1,0 +1,7 @@
+export default function getBundleNameFrom(component: string): string {
+  const bundleName = component.replace(/^@\//, '').replace(/\.[tj]sx?$/, '');
+  if (bundleName === 'pages/index') {
+    return 'host';
+  }
+  return bundleName;
+}
