+++ packages/umi-preset-react-native/src/plugins/generators/rn/templates/babelConfigTpl.ts
