+++ packages/umi-preset-react-native/src/plugins/generators/rn/templates/haulConfigTpl.ts
