+++ packages/umi-preset-react-native/src/plugins/generators/rn/templates/index.ts
